API/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.16
django-filter==22.1
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
django-graphql-social-auth==0.1.4
django-jsonfield-backport==1.0.5
django-rest-passwordreset==1.2.1
djongo==1.3.6
django-cors-headers==3.13.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.4
oauthlib==3.2.1
promise==2.3
pycparser==2.21
PyJWT==2.5.0
pymongo==3.12.3
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
text-unidecode==1.3
tzdata==2022.4
urllib3==1.26.12
